Introduce toenv for datastructures
toenv is like dump, but it exports the internal representation of
the datastructures.

The output format looks like this:

_T1='Foo'
_T2='Bar'
_L1='_T1
_T2'

Which is exactly how variables are spread in the shell process to
store each piece of data.

This helps not only troubleshooting, but fast save/restore of state
without having to re-interpret expressions and call constructors
all over again (although you might override stuff if the counters
are not properly accounted for).