Introduces the Map pseudodatastructure
The Map uses the Set for its list of keys, then hijacks the shell
variable namespace to store the mappings for each key.

Keys must be, therefore, valid variable names. This might change
in the future when we introduce hashing.

Changes were made to exp, dump and toenv to also work with it.