Makes _print and _write more familiar
_print and _write were weird because of our choice of IFS. When
passing multiple parameters to them, they would print one on each
line, which is not intuitive.

This changes them to accept a single parameter only, which makes
them less flexible, but much more predictable.

Also changed LC_ALL=C comment to better reflect our goal with
setting it. We're not after sort orders and stuff, what we want
is to operations like ${#var} (var length) to not consider multibyte
characters. I would LC_ALL=UTF-8 if all shells supported it, but
that isn't the case. Some will act as LC_ALL=C regardless of the
environment or changing the variable to a different locale string.