exp, Map, toenv, tap_tap usability, examples and test improvements
exp is now able to create Txt entries on the fly using ="value",
removing the need for intermediate user variables for it.

tap_tap had a bug that prevented failed tests with no output from
reporting, this was fixed

toenv was not properly serializing all references to a Map, this
was fixed

A mini how-to on working with datastructures is now on the examples
folder.