Fix Makefile call, drop ksh before 2012, drop posh 0.12
The docker-matrix being invoked in the Makefile was using "-c ./",
which caused the tested shell to invoke standard /bin/sh on each
run. This was causing all tests to be ran on the same shell, and
it was fixed.

Older ksh and posh were removed from the test matrix, since the
cost o keep them is to high (idiosyncrasies in parameters and
escapes).