Improving Lst tests
Using more of the idiom on the Lst tests, instead of doing derefs
by hand and comparing internal values.

Using dumps for assertions is a tradeoff. Tests that assert on
the internal representation would probably be better, but they're
not as good to read or maintain. Over time, as structures get
more stable, the effects of this tradeoff should be diminished.